+++ tests/user_test.py
@@ -1,5 +1,6 @@
 import unittest
 import sys
+from pathlib import Path
 from base_test_class import BaseTestCase
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
@@ -8,6 +9,18 @@ from selenium.webdriver import ActionChains
 
 
 class UserTest(BaseTestCase):
+
+    @staticmethod
+    def add_user_read_only_parameter():
+        f = open('/app/dojo/settings/local_settings.py', 'w')
+        f.write("USER_PROFILE_EDITABLE=False")
+        f.close()
+
+
+    @staticmethod
+    def reload_service():
+        Path("/app/dojo/settings/settings.py").touch()
+
 
     def test_create_user(self):
         # Login to the site.
@@ -43,6 +56,18 @@ class UserTest(BaseTestCase):
         # Assert ot the query to dtermine status of failure
         self.assertTrue(self.is_success_message_present(text='User added successfully, you may edit if necessary.') or
             self.is_success_message_present(text='A user with that username already exists.'))
+
+    def login_standard_page(self):
+        driver = self.driver
+        driver.get(self.base_url + "login")
+        driver.find_element_by_id("id_username").clear()
+        driver.find_element_by_id("id_username").send_keys('propersahm')
+        driver.find_element_by_id("id_password").clear()
+        driver.find_element_by_id("id_password").send_keys('Def3ctD0jo&')
+        driver.find_element_by_css_selector("button.btn.btn-success").click()
+
+        self.assertFalse(self.is_element_by_css_selector_present('.alert-danger', 'Please enter a correct username and password'))
+        return driver
 
     def test_user_edit_permissions(self):
         # Login to the site. Password will have to be modified
@@ -120,6 +145,18 @@ class UserTest(BaseTestCase):
         self.assertTrue(driver.find_element_by_xpath("//input[@name='product_added' and @value='mail']").is_selected())
         self.assertTrue(driver.find_element_by_xpath("//input[@name='scan_added' and @value='mail']").is_selected())
 
+    def test_standard_user_login(self):
+        self.login_standard_page()
+
+    def test_admin_profile_form(self):
+        self.add_user_read_only_parameter()
+        self.reload_service()
+        self.driver.get(self.base_url + "profile")
+        self.assertTrue(self.driver.find_element_by_id('id_first_name').is_enabled())
+
+    def test_user_profile_form(self):
+        self.driver.get(self.base_url + "profile")
+        self.assertFalse(self.driver.find_element_by_id('id_first_name').is_enabled())
 
 def suite():
     suite = unittest.TestSuite()
@@ -128,6 +165,12 @@ def suite():
     suite.addTest(BaseTestCase('test_login'))
     suite.addTest(UserTest('test_create_user'))
     suite.addTest(UserTest('test_user_edit_permissions'))
+    suite.addTest(UserTest('test_admin_profile_form'))
+    suite.addTest(BaseTestCase('test_logout'))
+    suite.addTest(UserTest('test_standard_user_login'))
+    suite.addTest(UserTest('test_user_profile_form'))
+    suite.addTest(BaseTestCase('test_logout'))
+    suite.addTest(BaseTestCase('test_login'))
     suite.addTest(UserTest('test_user_delete'))
 
     # not really for the user we created, but still related to user settings
